ekflogs: move log writer to seperate folder
JIRA: PP-169
